refactor(frontend/src/features/commands/dashboard
/FilterCmd.jsx):

move FilterCmd.jsx to features/commands/dashboard

for feature grouping
