+++ frontend/src/app/App.jsx
@@ -3,7 +3,7 @@ import { getCommand, getCommands } from "../features/commands/api/apiCommands";
 import { Navbar } from "./layout/Navbar";
 import { Route, Routes } from "react-router-dom";
 import { Home } from "../features/commands/lookup/Home";
-import { FilterCmd } from "../components/pages/Filter/FilterCmd";
+import { FilterCmd } from "../features/commands/dashboard/FilterCmd";
 import { CreateCmd } from "../components/pages/Filter/CreateCmd";
